+++ CHANGES.md
@@ -8,6 +8,8 @@ Changelog
 
 - Support for timeboxes started by users from Matrix rooms bridged to
   IRC channels using [NIMB][].
+- Command `summary` to display a summary of all timeboxes run across
+  all channels.
 
 ### Changed